-Fix (r1088): MSVC uint16 to uint8 warnings

+++ src/gui/gui.h
@@ -6,7 +6,7 @@
 extern void GUI_DrawWiredRectangle(uint16 left, uint16 top, uint16 right, uint16 bottom, uint16 colour);
 extern void GUI_DisplayText(const char *str, uint16 arg0A, ...);
 extern void GUI_DrawText(char *string, int16 left, int16 top, uint8 fgColour, uint8 bgColour);
-extern void GUI_DrawText_Wrapper(char *string, int16 left, int16 top, uint16 fgColour, uint16 bgColour, uint16 flags, ...);
+extern void GUI_DrawText_Wrapper(char *string, int16 left, int16 top, uint8 fgColour, uint8 bgColour, uint16 flags, ...);
 /* editbox.c */
 extern uint16 GUI_EditBox(csip32 text, uint16 maxLength, uint16 unknown1, csip32 wcsip, csip32 uknown3, uint16 unknown4);
